Cusum1.0data.json and Cusum1_0Bot(final).ipynb
The .json file represents the manmade query-response lists in "intents", whereas Cusum1_0Bot(final).ipynb is the Jupyter notebook in which I have written the code. users can interact with the bot in the bottommost cell in the notebook. 
Cusum1.0 is a Customer Support Bot for XYZ-GoShip (an imaginary company, of course) and responds to basic queries. 
This is a project submission for first use-case of Case1, as mentioned in the problem statement for the Capstone project, Summer Analytics 2024 by Consulting and Analytics Club, IIT Guwahati.